refactor(misc): add misc module to override
