+++ ui/src/components/admin/stats/Usages.vue
@@ -34,7 +34,7 @@
 </script>
 <script>
     import {mapStores} from "pinia";
-    import {useMiscStore} from "../../../stores/misc";
+    import {useMiscStore} from "override/stores/misc";
 
     export default {
         data() {
