+++ ui/src/components/layout/OnlyLeftMenuLayout.vue
@@ -10,7 +10,7 @@
     import LeftMenu from "override/components/LeftMenu.vue";
     import Errors from "../errors/Errors.vue";
     import {useCoreStore} from "../../stores/core";
-    import {useMiscStore} from "../../stores/misc";
+    import {useMiscStore} from "override/stores/misc";
     import {computed, onMounted} from "vue";
 
     const coreStore = useCoreStore();
