+++ ui/src/components/namespaces/components/NamespaceSelect.vue
@@ -25,7 +25,7 @@
 <script>
     import {mapState} from "vuex";
     import {mapStores} from "pinia";
-    import {useMiscStore} from "../../../stores/misc";
+    import {useMiscStore} from "override/stores/misc";
     import {useNamespacesStore} from "override/stores/namespaces";
     import _uniqBy from "lodash/uniqBy";
